// elearning-frontend/src/client-sdk/models/CourseProgressDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SectionProgressDto } from './SectionProgressDto';
export type CourseProgressDto = {
  sections: Array<SectionProgressDto>;
  id: string;
  title: string;
};